UI: Make the loop selector navigate to the loop on select
Change-type: patch
